Roll_tarot_html/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2 